# Add a faster code path for PETSc vectors.