--- a/tests/mpi/parallel_block_vector_04.mpirun=1.output
+++ b/tests/mpi/parallel_block_vector_04.mpirun=1.output
@@ -1,0 +1,12 @@
+
+DEAL:0::numproc=1
+DEAL:0::l2 norm: 3.464
+DEAL:0::l1 norm: 6.000
+DEAL:0::linfty norm: 2.000
+DEAL:0::l2.2 norm: 3.295
+DEAL:0::Mean value: 1.000
+DEAL:0::Inner product: 12.00
+DEAL:0:: v==0 ? 0
+DEAL:0:: v2==0 ? 1
+DEAL:0:: v2==0 ? 0
+DEAL:0::OK