--- a/tests/mpi/parallel_vector_19.mpirun=1.output
+++ b/tests/mpi/parallel_vector_19.mpirun=1.output
@@ -1,0 +1,3 @@
+
+DEAL:0::numproc=1
+DEAL:0::OK