--- a/tests/mpi/parallel_vector_19.mpirun=10.output
+++ b/tests/mpi/parallel_vector_19.mpirun=10.output
@@ -1,0 +1,3 @@
+
+DEAL:0::numproc=10
+DEAL:0::OK