--- a/tests/mpi/parallel_vector_19.mpirun=4.output
+++ b/tests/mpi/parallel_vector_19.mpirun=4.output
@@ -1,0 +1,3 @@
+
+DEAL:0::numproc=4
+DEAL:0::OK