@drwells 2016-12-28
It appears that some of the other vector classes have `add` (`Vector`, `TrilinosWrappers::MPI::Vector` both do) but our own `LinearAlgebra::distributed::Vector` does not. I suspect Trilinos vectors would also benefit from the same optimization: I'll try to generalize things.