@drwells 2016-12-28
I addressed the comments.

It looks like all we need to do to get rid of the PETSc specific bits is implement `add` for the `la_vector` classes, which I did. I marked this as WIP since those methods need tests. Is there a general design document describing what we expect what methods all vectors (PETSc, deal.II, etc) to implement? I don't know much about these classes.