@drwells 2016-12-29
I spent some time reading through #745 and related issues. Based on future plans to move things to inherit from `ReadWriteVector`, is it acceptable, for the sake of this patch, to implement `LinearAlgebra::Distributed::Vector::add` and `LinearAlgebra::Distributed::BlockVector::add` where I have? I believe that the answer is yes but I want to bring this up for explicit discussion.