@kronbichler 2016-12-29
> is it acceptable, for the sake of this patch, to implement LinearAlgebra::distributed::Vector::add and LinearAlgebra::distributed::BlockVector::add where I have? I believe that the answer is yes but I want to bring this up for explicit discussion.

I think this looks reasonable and do not see problems. Let's wait if @Rombur has comments, but in general the design of the vector classes was to start with a rather small interface and extend as necessary. This looks like a good approach of doing so. (Even though my codes do not have these kinds of code in the hot path so I cannot comment on possible alternatives - I guess for PETSc this is the way to go.)