@bangerth 2016-12-30
I didn't expect you to add these functions to the other vector classes, but I appreciate you did it anyway!

I'm in favor of merging as is. But Since @kronbichler wanted to see what @Rombur has to say, let's wait a couple of days to let him chime in.