@drwells 2016-12-30
I would still like to add a test for `distributed::BlockVector::add` and to rebase away my mistakes before we merge this.