@Rombur 2017-01-03
I think that this patch is fine but I don't think it should be extended to the new vector interface, i.e., it should not become part of the `VectorSpaceVector` interface. The `add(indices, values)` should be replaced by creating a  `ReadWriteVector rwv(indices)`, then filling the `ReadWriteVector`, and finally `vector.import(rwv, VectorOperation::add)`.