@drwells 2017-01-08
I cleaned things up and added the missing test.

@Rombur would you prefer that we do this interface change now or later? I am not sure that I understand what you are requesting.