@Rombur 2017-01-08
@drwells I am fine with the PR but I am against changing the interface in the new vector, for now. My message was more a reply to @kronbichler about extending the new vector interface.