@drwells 2017-01-09
@kronbichler I rewrote both descriptions: they now both explicitly describe the `add` method they use.